+++ sstcomp/PIMBackend/pim.cc
@@ -71,7 +71,7 @@ void TestPIM::read( Addr addr, uint64_t numBytes, std::vector<uint8_t>& payload 
     unsigned byte     = ( addr & 0x7 );
     assert( ( byte + numBytes ) <= 8 );  // 8 byte aligned only
     uint8_t* p = (uint8_t*) ( &( spdArray[spdIndex] ) );
-    for( int i = 0; i < numBytes; i++ ) {
+    for( unsigned i = 0; i < numBytes; i++ ) {
       payload[i] = p[byte + i];
     }
     output->verbose(
@@ -91,7 +91,7 @@ void TestPIM::write( Addr addr, uint64_t numBytes, std::vector<uint8_t>* payload
     unsigned byte   = ( addr & 0x7 );
     assert( ( byte + numBytes ) <= 8 );  // 8 byte aligned only
     uint8_t* p = (uint8_t*) ( &( spdArray[offset] ) );
-    for( int i = 0; i < numBytes; i++ ) {
+    for( unsigned i = 0; i < numBytes; i++ ) {
       p[byte + i] = payload->at( i );
     }
     output->verbose(
@@ -105,7 +105,7 @@ void TestPIM::write( Addr addr, uint64_t numBytes, std::vector<uint8_t>* payload
     assert( numBytes == 8 );
     assert( byte == 0 );
     uint8_t* p = (uint8_t*) ( &( ctl_buf ) );
-    for( int i = 0; i < numBytes; i++ ) {
+    for( unsigned i = 0; i < numBytes; i++ ) {
       p[byte + i] = payload->at( i );
     }
     output->verbose( CALL_INFO, 3, 0, "PIM 0x%" PRIx64 " IO WRITE FUNC A=0x%" PRIx64 " D=0x%" PRIx64 "\n", id, addr, ctl_buf );
@@ -168,7 +168,7 @@ unsigned TestPIM::SimpleDMA::clock() {
     parent->m_issueDRAMRequest( src, &parent->buffer, READ, [this]( const MemEventBase::dataVec& d ) 
       {
         assert( parent->buffer.size() == d.size() );
-        for( int i = 0; i < d.size(); i++ ) {
+        for( size_t i = 0; i < d.size(); i++ ) {
           parent->buffer[i] = d[i];
         }
         dma_state = DMA_STATE::WRITE;
